Services, Controllers, Caching

+++ Kubeauction/src/main/java/com/dmsrosa/kubeauction/database/dao/entity/AuctionEntity.java
@@ -34,7 +34,7 @@ public class AuctionEntity {
 
     private Date endDate;
 
-    private Double minimumPrice;
+    private Integer minimumPrice;
 
     @Builder.Default
     private Boolean isDeleted = false;
